requirements.txt
Django==5.2.7
psycopg==3.1.8
psycopg2
django-environ
python-dotenv==1.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
django-cors-headers==4.9.0
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.10.1

